feat(release): use pre-built Docker image and automate release of said image
Use release-please to automate releases of both the action and pre-built
Docker images, which are used to run the action, instead of having the
image dynamically built from a Dockerfile on each use.

This will require commits to follow Conventional Commit formatting going
forward, as it's a requirement of release-please. But doing so enables
automatic version bumping, changelog generation, and cutting new
releases by simply merging a "Release Pull Request" opened/updated by
release-please.

---

When this lands in master, release-please will run and raise a release
pull request with the title of this commit as a feature.

Due to action.yml being changed here to refer to a Docker image on
ghcr.io, it means the action will not work until the release PR is
merged, which bumps the image version and also builds and releases the
new image.